pyproject.toml
[tool.poetry]
name = "order-management"
version = "0.1.0"
description = ""
authors = ["Robert Jackiewicz <robert.jackiewicz@carta.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
django-rest-framework = "^0.1.0"
psycopg2 = "^2.9.3"
django-extensions = "^3.1.5"

[tool.poetry.dev-dependencies]
pylama = "^8.3.7"
black = "^22.1.0"
mypy = "^0.931"
django-stubs = "^1.9.0"
djangorestframework-stubs = "^1.4.0"
pre-commit = "^2.17.0"
pytest-django = "^4.5.2"
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
